Feature: Implement  Add and Remove functionality to To-Do List
