+++ .eslintignore
@@ -1,0 +1,2 @@
+dist/index.bundle.js
+dist/runtime.bundle.js
