+++ src/index.js
@@ -1,3 +1,3 @@
 import "./style.css";
-import "./modules/add.js"
+import "./modules/add_and_remove.js"
